CIL-1247 and CIL-1252 - Fix OIDC redirect problems
CIL-1247 OIDC clients are redirected to "Select an Identity Provider"
page when using Chrome. It's possible that CSRF cookie is being deleted
or overwritten. So for the /getuser/ and /secure/getuser/ endpoints,
don't check the CSRF cookie. Instead, just get the "submit" value.

CIL-1252 In the case that no 'responseurl' session variable has been
set, the code redirects the user to the "Select an Identity Provider"
page by default. Instead, check if either /device/ or /authorize/ (OIDC)
flows are in progress and attempt to redirect there instead.