CIL-1626 Redirect to OIDC client on common errors.
Before, the PHP code would catch errors from the
dbService?action=createTransaction call and display them to the end
user. However, many of those errors are due to incorrect OIDC query
parameters which should be returned to the OIDC client as documented at
https://www.rfc-editor.org/rfc/rfc6749#section-4.1.2.1 . The PHP code
now tries to catch these common errors and redirects the browser to the
redirect_uri with "error" and "error_description" parameters.