CIL-1812 Prevent open directs
When the PHP and OA4MP code bases were updated to handle better error
responses, that effort inadvertently enabled cilogon.org/authorize to be
an open redirector. This was due to the PHP code only requiring
'redirect_uri' to be passed in, and allowing the dbService to return a
valid error response. To fix this, make the PHP code check for all
required parameters before calling dbService?action=createTransaction.