CIL-1812 Allow session vars NOT to be output
Update the Loggit class to allow PHP session variables not to be output
to the syslogs. This is useful when a function calls info() multiple
times, but the session information hasn't changed. No point in filling
up the logs with duplicate debug output.

Also change several info() log messages to error() which makes more sense
for the context of that log message.

Also remove several XSEDE USAGE messages that got missed (CIL-1372).